requirements.txt
# Python 3.8
azure-functions==1.23.0b3
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
azure-storage-blob==12.19.0
azure-identity==1.15.0 